pyproject.toml
[project]
name = "config-control"
version = "0.1.0"
description = ""
authors = [
    {name = "askar",email = "askar.2033@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "twisted (>=25.5.0,<26.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "jsonschema (>=4.25.1,<5.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pytest (>=8.4.1,<9.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
